requirements.txt
git+https://github.com/HabanaAI/DeepSpeed.git@1.16.0
torch==2.3.1
pytorch_lightning==2.4.0
lightning_habana==1.6.0
